+++ Cargo.toml
@@ -4,7 +4,7 @@ description = "A tool for effective polyglot, multi-project monorepo development
 license = "MIT"
 homepage = "https://github.com/pnordahl/monorail"
 repository = "https://github.com/pnordahl/monorail"
-version = "3.0.4"
+version = "3.0.5"
 authors = ["Patrick Nordahl <plnordahl@gmail.com>"]
 edition = "2021"
 keywords = ["monorail", "monorepo", "build", "cli", "build-tool"]
