+++ src/api/cli.rs
@@ -1,5 +1,5 @@
-use crate::core;
-use crate::core::error::MonorailError;
+use crate::app;
+use crate::core::{self, error::MonorailError};
 
 use clap::builder::ArgPredicate;
 use clap::{Arg, ArgAction, ArgMatches, Command};
@@ -337,7 +337,7 @@ pub fn handle<'a>(
     output_options: &OutputOptions<'a>,
 ) -> Result<i32, MonorailError> {
     let verbosity = matches.get_one::<u8>(ARG_VERBOSE).unwrap_or(&0);
-    core::setup_tracing(output_options.format, *verbosity)?;
+    app::setup_tracing(output_options.format, *verbosity)?;
 
     match matches.get_one::<String>(ARG_CONFIG_FILE) {
         Some(config_file) => {
@@ -348,7 +348,7 @@ pub fn handle<'a>(
                     "Config file {} has no parent directory",
                     config_path.display()
                 )))?;
-            let config = core::app::Config::new(config_path)?;
+            let config = core::Config::new(config_path)?;
             if let Some(config_matches) = matches.subcommand_matches(CMD_CONFIG) {
                 if config_matches.subcommand_matches(CMD_SHOW).is_some() {
                     write_result(&Ok(config), output_options)?;
@@ -407,39 +407,39 @@ pub fn handle<'a>(
 }
 
 fn handle_out_delete<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
 ) -> Result<i32, MonorailError> {
-    let i = core::out::DeleteInput::try_from(matches)?;
-    let res = core::out::delete(&config.output_dir, &i);
+    let i = app::out::OutDeleteInput::try_from(matches)?;
+    let res = app::out::out_delete(&config.output_dir, &i);
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
 
 fn handle_analyze<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let i = core::HandleAnalyzeInput::from(matches);
-    let res = rt.block_on(core::handle_analyze(config, &i, work_path));
+    let i = app::analyze::HandleAnalyzeInput::from(matches);
+    let res = rt.block_on(app::analyze::handle_analyze(config, &i, work_path));
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
 
 fn handle_run<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let i = core::HandleRunInput::try_from(matches).unwrap();
+    let i = app::HandleRunInput::try_from(matches).unwrap();
     let invocation_args = env::args().skip(1).collect::<Vec<_>>().join(" ");
-    let o = rt.block_on(core::handle_run(config, &i, &invocation_args, work_path))?;
+    let o = rt.block_on(app::handle_run(config, &i, &invocation_args, work_path))?;
     let mut code = HANDLE_OK;
     if o.failed {
         code = HANDLE_ERR;
@@ -449,48 +449,48 @@ fn handle_run<'a>(
 }
 
 fn handle_checkpoint_update<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let i = core::CheckpointUpdateInput::try_from(matches)?;
-    let res = rt.block_on(core::handle_checkpoint_update(config, &i, work_path));
+    let i = app::CheckpointUpdateInput::try_from(matches)?;
+    let res = rt.block_on(app::handle_checkpoint_update(config, &i, work_path));
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
 
 fn handle_checkpoint_show<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let res = rt.block_on(core::handle_checkpoint_show(config, work_path));
+    let res = rt.block_on(app::handle_checkpoint_show(config, work_path));
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
 
 fn handle_checkpoint_delete<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let res = rt.block_on(core::handle_checkpoint_delete(config, work_path));
+    let res = rt.block_on(app::handle_checkpoint_delete(config, work_path));
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
 
 fn handle_log_tail<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
 ) -> Result<i32, MonorailError> {
     let rt = Runtime::new()?;
-    let i = core::log::LogTailInput::try_from(matches)?;
-    match rt.block_on(core::log::log_tail(config, &i)) {
+    let i = app::log::LogTailInput::try_from(matches)?;
+    match rt.block_on(app::log::log_tail(config, &i)) {
         Ok(_) => Ok(HANDLE_OK),
         Err(e) => {
             write_result(&Err::<(), MonorailError>(e), output_options)?;
@@ -500,13 +500,13 @@ fn handle_log_tail<'a>(
 }
 
 fn handle_log_show<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
-    let i = core::log::LogShowInput::try_from(matches)?;
-    match core::log::log_show(config, &i, work_path) {
+    let i = app::log::LogShowInput::try_from(matches)?;
+    match app::log::log_show(config, &i, work_path) {
         Ok(_) => Ok(HANDLE_OK),
         Err(e) => {
             write_result(&Err::<(), MonorailError>(e), output_options)?;
@@ -516,14 +516,14 @@ fn handle_log_show<'a>(
 }
 
 fn handle_target_list<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
-    let res = core::target_show(
+    let res = app::target_show(
         config,
-        core::TargetShowInput {
+        app::TargetShowInput {
             show_target_groups: matches.get_flag(ARG_TARGET_GROUPS),
         },
         work_path,
@@ -532,13 +532,13 @@ fn handle_target_list<'a>(
     Ok(get_code(res.is_err()))
 }
 fn handle_result_show<'a>(
-    config: &'a core::app::Config,
+    config: &'a core::Config,
     matches: &'a ArgMatches,
     output_options: &OutputOptions<'a>,
     work_path: &'a path::Path,
 ) -> Result<i32, MonorailError> {
-    let i = core::ResultShowInput::try_from(matches)?;
-    let res = core::result_show(config, work_path, &i);
+    let i = app::ResultShowInput::try_from(matches)?;
+    let res = app::result_show(config, work_path, &i);
     write_result(&res, output_options)?;
     Ok(get_code(res.is_err()))
 }
@@ -588,19 +588,19 @@ where
     }
 }
 
-impl TryFrom<&clap::ArgMatches> for core::ResultShowInput {
+impl TryFrom<&clap::ArgMatches> for app::ResultShowInput {
     type Error = MonorailError;
     fn try_from(_cmd: &clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {})
     }
 }
 
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::CheckpointUpdateInput<'a> {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::CheckpointUpdateInput<'a> {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
             id: cmd.get_one::<String>(ARG_ID).map(|x: &String| x.as_str()),
-            git_opts: core::GitOptions {
+            git_opts: core::git::GitOptions {
                 start: None,
                 end: None,
                 git_path: cmd
@@ -611,11 +611,11 @@ impl<'a> TryFrom<&'a clap::ArgMatches> for core::CheckpointUpdateInput<'a> {
         })
     }
 }
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::HandleRunInput<'a> {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::HandleRunInput<'a> {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
-            git_opts: core::GitOptions {
+            git_opts: core::git::GitOptions {
                 start: cmd
                     .get_one::<String>(ARG_START)
                     .map(|x: &String| x.as_str()),
@@ -641,10 +641,10 @@ impl<'a> TryFrom<&'a clap::ArgMatches> for core::HandleRunInput<'a> {
     }
 }
 
-impl<'a> From<&'a clap::ArgMatches> for core::HandleAnalyzeInput<'a> {
+impl<'a> From<&'a clap::ArgMatches> for app::analyze::HandleAnalyzeInput<'a> {
     fn from(cmd: &'a clap::ArgMatches) -> Self {
         Self {
-            git_opts: core::GitOptions {
+            git_opts: core::git::GitOptions {
                 start: cmd
                     .get_one::<String>(ARG_START)
                     .map(|x: &String| x.as_str()),
@@ -656,7 +656,7 @@ impl<'a> From<&'a clap::ArgMatches> for core::HandleAnalyzeInput<'a> {
                 .into_iter()
                 .flatten()
                 .collect(),
-            analyze_input: core::AnalyzeInput {
+            analyze_input: app::analyze::AnalyzeInput {
                 show_changes: cmd.get_flag(ARG_CHANGES),
                 show_change_targets: cmd.get_flag(ARG_CHANGE_TARGETS),
                 show_target_groups: cmd.get_flag(ARG_TARGET_GROUPS),
@@ -665,16 +665,16 @@ impl<'a> From<&'a clap::ArgMatches> for core::HandleAnalyzeInput<'a> {
     }
 }
 
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::log::LogTailInput {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::log::LogTailInput {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
-            filter_input: core::log::FilterInput::try_from(cmd)?,
+            filter_input: app::log::LogFilterInput::try_from(cmd)?,
         })
     }
 }
 
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::log::FilterInput {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::log::LogFilterInput {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
@@ -696,17 +696,17 @@ impl<'a> TryFrom<&'a clap::ArgMatches> for core::log::FilterInput {
     }
 }
 
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::log::LogShowInput<'a> {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::log::LogShowInput<'a> {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
-            filter_input: core::log::FilterInput::try_from(cmd)?,
+            filter_input: app::log::LogFilterInput::try_from(cmd)?,
             id: cmd.get_one::<usize>(ARG_ID),
         })
     }
 }
 
-impl<'a> TryFrom<&'a clap::ArgMatches> for core::out::DeleteInput {
+impl<'a> TryFrom<&'a clap::ArgMatches> for app::out::OutDeleteInput {
     type Error = MonorailError;
     fn try_from(cmd: &'a clap::ArgMatches) -> Result<Self, Self::Error> {
         Ok(Self {
