+++ src/app/log.rs
@@ -10,28 +10,14 @@ use std::io::BufRead;
 use tokio::io::{AsyncBufReadExt, AsyncWriteExt};
 use tracing::{debug, info, instrument, trace};
 
-use crate::core::{app, error::MonorailError, tracking};
+use crate::core::{self, error::MonorailError, tracking};
 
 pub(crate) const STDOUT_FILE: &str = "stdout.zst";
 pub(crate) const STDERR_FILE: &str = "stderr.zst";
 pub(crate) const RESET_COLOR: &str = "\x1b[0m";
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub(crate) struct Config {
-    // Tick frequency for flushing accumulated logs to stream
-    // and compression tasks
-    flush_interval_ms: u64,
-}
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            flush_interval_ms: 500,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub(crate) struct FilterInput {
+pub(crate) struct LogFilterInput {
     pub(crate) commands: HashSet<String>,
     pub(crate) targets: HashSet<String>,
     pub(crate) include_stdout: bool,
@@ -40,11 +26,11 @@ pub(crate) struct FilterInput {
 
 #[derive(Serialize)]
 pub(crate) struct LogTailInput {
-    pub(crate) filter_input: FilterInput,
+    pub(crate) filter_input: LogFilterInput,
 }
 
 pub(crate) async fn log_tail<'a>(
-    _cfg: &'a app::Config,
+    _cfg: &'a core::Config,
     input: &LogTailInput,
 ) -> Result<(), MonorailError> {
     let mut lss = StreamServer::new("127.0.0.1:9201", &input.filter_input);
@@ -54,14 +40,20 @@ pub(crate) async fn log_tail<'a>(
 #[derive(Serialize)]
 pub(crate) struct LogShowInput<'a> {
     pub(crate) id: Option<&'a usize>,
-    pub(crate) filter_input: FilterInput,
+    pub(crate) filter_input: LogFilterInput,
 }
 
 pub(crate) fn log_show<'a>(
-    cfg: &'a app::Config,
+    cfg: &'a core::Config,
     input: &'a LogShowInput<'a>,
     work_path: &'a path::Path,
 ) -> Result<(), MonorailError> {
+    // require at least one of the log types be opted into
+    if !input.filter_input.include_stdout && !input.filter_input.include_stderr {
+        return Err(MonorailError::from(
+            "No stream selected; provide one or both of: --stdout, --stderr",
+        ));
+    }
     let log_id = match input.id {
         Some(id) => *id,
         None => {
@@ -85,6 +77,7 @@ pub(crate) fn log_show<'a>(
             "No configured targets, cannot tail logs",
         ));
     }
+
     let mut hasher = sha2::Sha256::new();
     let mut hash2target = HashMap::new();
     for target in &cfg.targets {
@@ -176,10 +169,10 @@ fn stream_archive_file_to_stdout(
 
 pub(crate) struct StreamServer<'a> {
     address: &'a str,
-    args: &'a FilterInput,
+    args: &'a LogFilterInput,
 }
 impl<'a> StreamServer<'a> {
-    pub(crate) fn new(address: &'a str, args: &'a FilterInput) -> Self {
+    pub(crate) fn new(address: &'a str, args: &'a LogFilterInput) -> Self {
         Self { address, args }
     }
     pub(crate) async fn listen(&mut self) -> Result<(), MonorailError> {
@@ -213,7 +206,7 @@ impl<'a> StreamServer<'a> {
 #[derive(Debug, Clone)]
 pub(crate) struct StreamClient {
     stream: sync::Arc<tokio::sync::Mutex<tokio::net::TcpStream>>,
-    pub(crate) args: FilterInput,
+    pub(crate) args: LogFilterInput,
 }
 impl StreamClient {
     #[instrument]
@@ -237,7 +230,7 @@ impl StreamClient {
         // pull arg preferences from the server on connect
         let mut br = tokio::io::BufReader::new(&mut stream);
         br.read_until(b'\n', &mut args_data).await?;
-        let args: FilterInput = serde_json::from_slice(args_data.as_slice())?;
+        let args: LogFilterInput = serde_json::from_slice(args_data.as_slice())?;
         debug!("Received log stream arguments");
         if args.include_stdout || args.include_stderr {
             let targets = if args.targets.is_empty() {
@@ -271,7 +264,7 @@ impl StreamClient {
 }
 
 pub(crate) async fn process_reader<R>(
-    config: &Config,
+    config: &core::LogConfig,
     mut reader: tokio::io::BufReader<R>,
     compressor_client: CompressorClient,
     header: String,
