+++ src/core/testing.rs
@@ -1,20 +1,43 @@
-use std::sync::atomic::{AtomicUsize, Ordering};
+use crate::core;
+use sha2::Digest;
+use std::collections::HashMap;
 use std::path;
+use std::sync::atomic::{AtomicUsize, Ordering};
 use tokio::io::AsyncWriteExt;
+
+pub(crate) const RAW_CONFIG: &str = r#"
+{
+    "targets": [
+        {
+            "path": "rust"
+        },
+        {
+            "path": "rust/target",
+            "ignores": [
+                "rust/target/ignoreme.txt"
+            ],
+            "uses": [
+                "rust/vendor",
+                "common"
+            ]
+        }
+    ]
+}
+"#;
 
 // Using this in tests allows them to execute concurrently with a clean repo in each case.
 static GLOBAL_REPO_ID: AtomicUsize = AtomicUsize::new(0);
 
 // Git utils
-pub fn test_path() -> path::PathBuf {
+pub(crate) fn test_path() -> path::PathBuf {
     path::Path::new(&"/tmp/monorail".to_string()).to_path_buf()
 }
 
-pub fn repo_path(test_path: &path::Path, n: usize) -> path::PathBuf {
+pub(crate) fn repo_path(test_path: &path::Path, n: usize) -> path::PathBuf {
     test_path.join(format!("test-{}", n))
 }
 
-pub async fn init(bare: bool) -> path::PathBuf {
+pub(crate) async fn init(bare: bool) -> path::PathBuf {
     let id = GLOBAL_REPO_ID.fetch_add(1, Ordering::SeqCst);
     let test_path = test_path();
     tokio::fs::create_dir_all(&test_path).await.unwrap();
@@ -35,7 +58,7 @@ pub async fn init(bare: bool) -> path::PathBuf {
     repo_path
 }
 
-pub async fn add(name: &str, repo_path: &path::Path) {
+pub(crate) async fn add(name: &str, repo_path: &path::Path) {
     let _ = tokio::process::Command::new("git")
         .arg("add")
         .arg(name)
@@ -44,7 +67,7 @@ pub async fn add(name: &str, repo_path: &path::Path) {
         .await
         .unwrap();
 }
-pub async fn commit(repo_path: &path::Path) {
+pub(crate) async fn commit(repo_path: &path::Path) {
     let _ = tokio::process::Command::new("git")
         .arg("commit")
         .arg("-a")
@@ -56,7 +79,7 @@ pub async fn commit(repo_path: &path::Path) {
         .await
         .unwrap();
 }
-pub async fn get_head(repo_path: &path::Path) -> String {
+pub(crate) async fn get_head(repo_path: &path::Path) -> String {
     let end = String::from_utf8(
         tokio::process::Command::new("git")
             .arg("rev-parse")
@@ -70,7 +93,7 @@ pub async fn get_head(repo_path: &path::Path) -> String {
     .unwrap();
     String::from(end.trim())
 }
-pub async fn create_file(
+pub(crate) async fn create_file(
     repo_path: &path::Path,
     dir: &str,
     file_name: &str,
@@ -87,4 +110,43 @@ pub async fn create_file(
         .unwrap();
     file.write_all(content).await.unwrap();
     file
+}
+
+pub(crate) async fn write_with_checksum(
+    path: &path::Path,
+    data: &[u8],
+) -> Result<String, tokio::io::Error> {
+    let mut hasher = sha2::Sha256::new();
+    hasher.update(data);
+    tokio::fs::write(path, &data).await?;
+    Ok(hex::encode(hasher.finalize()).to_string())
+}
+
+pub(crate) fn get_pair_map(pairs: &[(&str, String)]) -> HashMap<String, String> {
+    let mut pending = HashMap::new();
+    for (fname, checksum) in pairs {
+        pending.insert(fname.to_string(), checksum.clone());
+    }
+    pending
+}
+
+pub(crate) async fn prep_raw_config_repo() -> (core::Config, path::PathBuf) {
+    let repo_path = init(false).await;
+    let c: core::Config = serde_json::from_str(RAW_CONFIG).unwrap();
+
+    create_file(
+        &repo_path,
+        "rust",
+        "monorail.sh",
+        b"command whoami { echo 'rust' }",
+    )
+    .await;
+    create_file(
+        &repo_path,
+        "rust/target",
+        "monorail.sh",
+        b"command whoami { echo 'rust/target' }",
+    )
+    .await;
+    (c, repo_path)
 }
