+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod api;
+pub mod app;
 pub mod core;
